feat: refactor repository structure and cleanup
