+++ .devcontainer/docker-compose.yml
@@ -20,60 +20,6 @@ services:
       - PATH=/opt/venv/bin:$PATH
       - PYTHONDONTWRITEBYTECODE=1
       - PYTHONUNBUFFERED=1
-    depends_on:
-      - redis
-      - mysql
-
-  dify:
-    build:
-      context: ./dify/docker
-      dockerfile: Dockerfile
-    volumes:
-      - dify_data:/app/data
-    ports:
-      - "3000:3000"
-    environment:
-      - NODE_ENV=development
-      - CONSOLE_URL=http://localhost:3000
-      - API_URL=http://localhost:5001
-      - APP_URL=http://localhost:3000
-      - STORAGE_TYPE=local
-      - STORAGE_LOCAL_PATH=/app/data
-      - REDIS_HOST=redis
-      - REDIS_PORT=6379
-      - REDIS_USERNAME=
-      - REDIS_PASSWORD=
-      - REDIS_DB=0
-      - MYSQL_HOST=mysql
-      - MYSQL_PORT=3306
-      - MYSQL_USERNAME=root
-      - MYSQL_PASSWORD=password
-      - MYSQL_DATABASE=dify
-    depends_on:
-      - redis
-      - mysql
-
-  redis:
-    image: redis:7.0
-    ports:
-      - "6379:6379"
-    volumes:
-      - redis_data:/data
-    command: redis-server --appendonly yes
-
-  mysql:
-    image: mysql:8.0
-    ports:
-      - "3306:3306"
-    volumes:
-      - mysql_data:/var/lib/mysql
-    environment:
-      - MYSQL_ROOT_PASSWORD=password
-      - MYSQL_DATABASE=dify
-    command: --default-authentication-plugin=mysql_native_password
 
 volumes:
   venv:
-  dify_data:
-  redis_data:
-  mysql_data:
