--- Dockerfile
@@ -1,15 +1,0 @@
-FROM python:3.9-slim
-
-WORKDIR /app
-
-# Install curl
-RUN apt-get update && apt-get install -y curl && rm -rf /var/lib/apt/lists/*
-
-COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY . .
-
-EXPOSE 8000
-
-CMD ["uvicorn", "jira_api:app", "--host", "0.0.0.0", "--port", "8000"]
