+++ config/.env.example
@@ -1,0 +1,8 @@
+# Jira credentials
+JIRA_SERVER_URL=https://your-domain.atlassian.net
+JIRA_EMAIL=your-email@example.com
+JIRA_API_TOKEN=your-api-token
+
+# Dify credentials
+DIFY_DATASET_API_KEY=your-dify-api-key
+DIFY_BASE_URL=https://api.dify.ai/v1 # Optional, defaults to this value
