--- pyproject.toml
@@ -1,54 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "dify-jira"
-version = "0.1.0"
-description = "Integration between Jira and Dify RAG system"
-readme = "README.md"
-requires-python = ">=3.9"
-license = "MIT"
-authors = [
-    { name = "Your Name", email = "your.email@example.com" }
-]
-dependencies = [
-    "jira==3.5.2",
-    "python-dotenv==1.0.0",
-    "requests==2.31.0",
-    "urllib3==2.0.7",
-    "pydantic==2.5.2"
-]
-
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0",
-    "pytest-cov>=4.0",
-    "black>=23.0",
-    "isort>=5.0",
-    "mypy>=1.0",
-    "pylint>=3.0",
-    "pre-commit>=3.0"
-]
-
-[tool.black]
-line-length = 88
-target-version = ["py39"]
-include = '\.pyi?$'
-
-[tool.isort]
-profile = "black"
-multi_line_output = 3
-line_length = 88
-
-[tool.mypy]
-python_version = "3.9"
-warn_return_any = true
-warn_unused_configs = true
-disallow_untyped_defs = true
-disallow_incomplete_defs = true
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-python_files = ["test_*.py"]
-addopts = "--cov=jira_rag --cov-report=term-missing"
