+++ scripts/setup-dify.sh
