+++ src/core/models/__init__.py
@@ -1,0 +1,3 @@
+from .ingest_models import IngestJiraRequest, IngestJsonRequest
+
+__all__ = ['IngestJiraRequest', 'IngestJsonRequest']
